# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EPIS2 Clinical API",
    "description": "API modular para ficha clinica inteligente y auditable.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List Patients",
        "operationId": "list_patients_api_v1_patients_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-EPIS2-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Epis2-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientRead"
                  },
                  "title": "Response List Patients Api V1 Patients Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Patient",
        "operationId": "create_patient_api_v1_patients_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-EPIS2-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Epis2-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patient",
        "operationId": "get_patient_api_v1_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-EPIS2-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Epis2-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "patients"
        ],
        "summary": "Update Patient",
        "operationId": "update_patient_api_v1_patients__patient_id__patch",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-EPIS2-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Epis2-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/record": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patient Record",
        "operationId": "get_patient_record_api_v1_patients__patient_id__record_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-EPIS2-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Epis2-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRecordSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/ai/suggestions": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Patient Ai Suggestions",
        "operationId": "create_patient_ai_suggestions_api_v1_patients__patient_id__ai_suggestions_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-EPIS2-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Epis2-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientAiSuggestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientAiSuggestionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AllergyRead": {
        "properties": {
          "substance": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Substance"
          },
          "reaction": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Reaction"
          },
          "severity": {
            "$ref": "#/components/schemas/AllergySeverity",
            "default": "unknown"
          },
          "status": {
            "$ref": "#/components/schemas/RecordStatus",
            "default": "active"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "substance",
          "recorded_at",
          "id",
          "patient_id",
          "created_at",
          "updated_at"
        ],
        "title": "AllergyRead"
      },
      "AllergySeverity": {
        "type": "string",
        "enum": [
          "mild",
          "moderate",
          "severe",
          "unknown"
        ],
        "title": "AllergySeverity"
      },
...
      "CareContext": {
        "type": "string",
        "enum": [
          "ambulatory",
          "hospitalized",
          "unknown"
        ],
        "title": "CareContext"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientAiSuggestionRequest": {
        "properties": {
          "focus": {
            "type": "string",
            "enum": [
              "summary",
              "safety",
              "documentation"
            ],
            "title": "Focus",
            "default": "summary"
          }
        },
        "type": "object",
        "title": "PatientAiSuggestionRequest"
      },
      "PatientAiSuggestionsResponse": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "enum": [
              "disabled",
              "draft",
              "error"
            ],
            "title": "Status"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/ClinicalAiSuggestion"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Safety Notes"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "status",
          "patient_id",
          "generated_at",
          "summary"
        ],
        "title": "PatientAiSuggestionsResponse"
      },
      "PatientClinicalStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "closed",
          "archived"
        ],
        "title": "PatientClinicalStatus"
      },
      "PatientCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "sex_at_birth": {
            "$ref": "#/components/schemas/SexAtBirth",
            "default": "unknown"
          },
          "clinical_status": {
            "$ref": "#/components/schemas/PatientClinicalStatus",
            "default": "active"
          },
          "current_care_context": {
            "$ref": "#/components/schemas/CareContext",
            "default": "unknown"
          },
          "document_id_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id Hash"
          },
          "clinical_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Identifier"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "emergency_contact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Emergency Contact"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "birth_date"
        ],
        "title": "PatientCreate"
      },
      "PatientRead": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "sex_at_birth": {
            "$ref": "#/components/schemas/SexAtBirth",
            "default": "unknown"
          },
          "clinical_status": {
            "$ref": "#/components/schemas/PatientClinicalStatus",
            "default": "active"
          },
          "current_care_context": {
            "$ref": "#/components/schemas/CareContext",
            "default": "unknown"
          },
          "document_id_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id Hash"
          },
          "clinical_identifier": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Identifier"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "emergency_contact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Emergency Contact"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "birth_date",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "PatientRead"
      },
      "PatientRecordSnapshot": {
        "properties": {
          "patient": {
            "$ref": "#/components/schemas/PatientRead"
          },
          "latest_vitals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalSignRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyRead"
            },
            "type": "array",
            "title": "Active Allergies"
          },
          "active_medications": {
            "items": {
              "$ref": "#/components/schemas/MedicationRead"
            },
            "type": "array",
            "title": "Active Medications"
          },
          "active_problems": {
            "items": {
              "$ref": "#/components/schemas/ActiveProblemRead"
            },
            "type": "array",
            "title": "Active Problems"
          },
          "recent_entries": {
            "items": {
              "$ref": "#/components/schemas/ClinicalEntryRead"
            },
            "type": "array",
            "title": "Recent Entries"
          }
        },
        "type": "object",
        "required": [
          "patient"
        ],
        "title": "PatientRecordSnapshot"
      },
      "PatientUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "emergency_contact": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact"
          },
          "clinical_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientClinicalStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_care_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CareContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "PatientUpdate"
      },
      "RecordStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "resolved",
          "entered_in_error"
        ],
        "title": "RecordStatus"
      },
      "SexAtBirth": {
        "type": "string",
        "enum": [
          "female",
          "male",
          "intersex",
          "unknown"
        ],
        "title": "SexAtBirth"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VitalSignRead": {
        "properties": {
          "measured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Measured At"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "systolic_bp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300.0,
                "minimum": 40.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Systolic Bp"
          },
          "diastolic_bp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Diastolic Bp"
          },
          "heart_rate_bpm": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 260.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Bpm"
          },
          "respiratory_rate_bpm": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 80.0,
                "minimum": 4.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Respiratory Rate Bpm"
          },
          "oxygen_saturation_pct": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oxygen Saturation Pct"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "measured_at",
          "id",
          "patient_id",
          "created_at",
          "updated_at"
        ],
        "title": "VitalSignRead"
      },
...
    }
  }
}